+++ src/main/java/de/blablubbabc/shopkeepers/bluemap/ShopkeepersBlueMap.java
@@ -2,6 +2,8 @@ package de.blablubbabc.shopkeepers.bluemap;
 
 import java.io.IOException;
 import java.nio.file.Files;
+import java.util.concurrent.locks.ReentrantLock;
+import java.util.function.Consumer;
 import java.util.logging.Level;
 
 import org.bukkit.Bukkit;
@@ -12,12 +14,17 @@ import com.nisovin.shopkeepers.api.ShopkeepersAPI;
 import com.nisovin.shopkeepers.api.shopkeeper.Shopkeeper;
 import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopkeeper;
 
+import de.blablubbabc.shopkeepers.bluemap.util.SchedulerUtils;
+
 import de.bluecolored.bluemap.api.BlueMapAPI;
 import de.bluecolored.bluemap.api.BlueMapMap;
 import de.bluecolored.bluemap.api.BlueMapWorld;
 import de.bluecolored.bluemap.api.markers.MarkerSet;
 import de.bluecolored.bluemap.api.markers.POIMarker;
 
+/**
+ * Shopkeepers BlueMap integration.
+ */
 public class ShopkeepersBlueMap {
 
 	private static final String WEB_SHOPKEEPERS_ASSETS = "assets/shopkeepers";
@@ -45,14 +52,24 @@ public class ShopkeepersBlueMap {
 	private final ShopkeepersBlueMapPlugin plugin;
 	private final ShopkeepersListener shopkeeperListener = new ShopkeepersListener(this);
 
-	private boolean enabled = false;
-
+	// Fair reentrant lock to ensure consistent ordering of enable and disable callbacks:
+	private final ReentrantLock blueMapLock = new ReentrantLock(true);
+	// Do not access this except while holding the blueMapLock (see runWithBlueMapLock and
+	// runBlueMapOperation):
 	private BlueMapAPI blueMapApi = null;
+
+	private boolean enabled = false;
+	private boolean assetsWritten = false;
 
 	public ShopkeepersBlueMap(ShopkeepersBlueMapPlugin plugin) {
 		this.plugin = plugin;
 	}
 
+	/**
+	 * Enables the integration.
+	 * <p>
+	 * For example called during plugin enable.
+	 */
 	public void enable() {
 		if (enabled) {
 			return;
@@ -63,6 +80,8 @@ public class ShopkeepersBlueMap {
 		}
 
 		enabled = true;
+		// Try to write the assets again after each reload:
+		assetsWritten = false;
 
 		// Called immediately if the BlueMap API is currently enabled:
 		BlueMapAPI.onEnable(this::onBlueMapEnabledAsync);
@@ -71,6 +90,11 @@ public class ShopkeepersBlueMap {
 		Bukkit.getPluginManager().registerEvents(shopkeeperListener, plugin);
 	}
 
+	/**
+	 * Disables the integration.
+	 * <p>
+	 * For example called during plugin disable.
+	 */
 	public void disable() {
 		if (!enabled) {
 			return;
@@ -81,75 +105,130 @@ public class ShopkeepersBlueMap {
 		BlueMapAPI.unregisterListener(this::onBlueMapEnabledAsync);
 		BlueMapAPI.unregisterListener(this::onBlueMapDisabledAsync);
 
-		this.removeAllShopkeepers();
+		this.onBlueMapDisabledAsync();
 
 		enabled = false;
 	}
 
-	// According to the documentation, this may be called off the main server thread!
-	private void onBlueMapEnabledAsync(BlueMapAPI bluemap) {
-		if (!plugin.isEnabled()) {
-			return;
+	/**
+	 * Synchronizes access to {@link ShopkeepersBlueMap#blueMapApi} to only allow one operation at a
+	 * time.
+	 * <p>
+	 * This is used to resolve race conditions. For example, when the BlueMap API is disabled off
+	 * the main thread, we wait for any in-progress marker updates on the main thread to complete
+	 * before we cleanup all markers. Otherwise, we could end up cleaning up the markers off the
+	 * main thread just for new markers to be subsequently added on the main thread.
+	 * <p>
+	 * The operations are run in FIFO order.
+	 * 
+	 * @param operation
+	 *            the operation to run inside the BlueMap lock
+	 */
+	private void runWithBlueMapLock(Runnable operation) {
+		blueMapLock.lock();
+		try {
+			operation.run();
+		} finally {
+			blueMapLock.unlock();
 		}
+	}
 
-		if (Bukkit.isPrimaryThread()) {
-			this.onBlueMapEnabled(bluemap);
-			return;
-		}
+	/**
+	 * See {@link #runWithBlueMapLock(Runnable)}.
+	 * <p>
+	 * The operation is skipped if the BlueMap API is not available currently.
+	 * 
+	 * @param operation
+	 *            the BlueMap operation to run
+	 */
+	private void runBlueMapOperation(Consumer<BlueMapAPI> operation) {
+		this.runWithBlueMapLock(() -> {
+			var blueMapApi = this.blueMapApi;
+			if (blueMapApi == null) {
+				return;
+			}
 
-		Bukkit.getScheduler().runTask(plugin, () -> {
-			BlueMapAPI.getInstance().ifPresent(this::onBlueMapEnabled);
+			operation.accept(blueMapApi);
 		});
 	}
 
 	// According to the documentation, this may be called off the main server thread!
-	private void onBlueMapDisabledAsync(BlueMapAPI bluemap) {
-		if (!plugin.isEnabled()) {
-			return;
-		}
-
-		if (Bukkit.isPrimaryThread()) {
-			this.onBlueMapDisabled(bluemap);
-			return;
-		}
-
-		Bukkit.getScheduler().runTask(plugin, () -> {
-			BlueMapAPI.getInstance().ifPresent(this::onBlueMapDisabled);
+	private void onBlueMapEnabledAsync(BlueMapAPI blueMapApi) {
+		SchedulerUtils.runOnMainThreadOrOmit(plugin, () -> {
+			this.runWithBlueMapLock(() -> {
+				// Only enable if the BlueMap API is still enabled:
+				BlueMapAPI.getInstance().ifPresent(this::onBlueMapEnabled);
+			});
 		});
 	}
 
-	private void onBlueMapEnabled(BlueMapAPI bluemap) {
-		if (!enabled) {
-			return;
-		}
-
-		if (this.blueMapApi != null) {
-			return;
-		}
-
-		this.blueMapApi = bluemap;
-
-		this.writeAssets();
-		this.addAllShopkeepers();
+	// According to the documentation, this may be called off the main server thread!
+	private void onBlueMapDisabledAsync(BlueMapAPI _unused) {
+		// Note: BlueMapAPI.getInstance is already null at this point.
+		// Note: We handle the cleanup immediately, potentially off the main thread, because it is
+		// not guaranteed that the API can still be used after this method returns (e.g. after
+		// synchronization to the main thread).
+		// This intentionally blocks while waiting for other BlueMap operations to complete.
+		this.runWithBlueMapLock(() -> {
+			this.onBlueMapDisabledAsync();
+		});
 	}
 
-	private void onBlueMapDisabled(BlueMapAPI _unused) {
+	// Called on the main thread inside the BlueMap lock.
+	private void onBlueMapEnabled(BlueMapAPI newBlueMapApi) {
 		if (!enabled) {
 			return;
 		}
+		// This is called on the main thread and we disable the integration during plugin disable:
+		assert plugin.isEnabled();
 
 		if (this.blueMapApi != null) {
+			// Already enabled.
+			// Unexpected, because API enable and disable callbacks are run in FIFO order and
+			// disable blocks while waiting for the lock.
 			return;
 		}
 
-		this.removeAllShopkeepers();
+		this.blueMapApi = newBlueMapApi;
+
+		if (!assetsWritten) {
+			assetsWritten = true;
+
+			// Write the assets asynchronously and add the shopkeeper markers afterwards (on the
+			// main thread):
+			SchedulerUtils.runAsyncTaskOrOmit(plugin, () -> {
+				this.runBlueMapOperation(blueMapApi -> {
+					this.writeAssets(blueMapApi);
+				});
+
+				SchedulerUtils.runOnMainThreadOrOmit(plugin, () -> {
+					this.runBlueMapOperation(blueMapApi -> {
+						this.addAllShopkeepers(blueMapApi);
+					});
+				});
+			});
+			return;
+		}
+
+		this.addAllShopkeepers(newBlueMapApi);
+	}
+
+	// Potentially called off the main thread, while holding the BlueMap lock.
+	private void onBlueMapDisabledAsync() {
+		var blueMapApi = this.blueMapApi;
+		if (blueMapApi == null) {
+			return; // Already disabled
+		}
+		// blueMapApi is only assigned when the integration is enabled:
+		assert enabled;
+
+		this.removeAllShopkeepersAsync(blueMapApi);
 
 		this.blueMapApi = null;
 	}
 
-	private void writeAssets() {
-		assert this.enabled && this.blueMapApi != null;
-
+	// Potentially called off the main thread.
+	private void writeAssets(BlueMapAPI blueMapApi) {
 		plugin.getLogger().info("Writing web assets.");
 
 		// The marker icons are the same across all worlds and maps. We therefore save them to the
@@ -191,28 +270,63 @@ public class ShopkeepersBlueMap {
 		}
 	}
 
-	private void addAllShopkeepers() {
-		// Note: If the Shopkeepers API is later enabled, the shopkeepers will be added one-by-one
-		// via the ShopkeeperAddedEvent.
-		if (!ShopkeepersAPI.isEnabled() || this.blueMapApi == null) {
-			return;
-		}
-
-		ShopkeepersAPI.getShopkeeperRegistry().getAllShopkeepers().forEach(this::addShopkeeper);
+	private MarkerSet getOrCreateMarkerSet(BlueMapMap map) {
+		return map.getMarkerSets().computeIfAbsent(
+				MARKERSET_ID,
+				key -> MarkerSet.builder()
+						.label(plugin.getSettings().getMarkerSetName())
+						.defaultHidden(false)
+						.toggleable(true)
+						.build()
+		);
 	}
 
-	private void removeAllShopkeepers() {
-		if (!ShopkeepersAPI.isEnabled() || this.blueMapApi == null) {
+	private MarkerSet getMarkerSet(BlueMapMap map) {
+		return map.getMarkerSets().get(MARKERSET_ID);
+	}
+
+	private MarkerSet removeMarkerSet(BlueMapMap map) {
+		return map.getMarkerSets().remove(MARKERSET_ID);
+	}
+
+	private void addAllShopkeepers(BlueMapAPI blueMapApi) {
+		// Note: If the Shopkeepers API is later enabled, the shopkeepers will be added one-by-one
+		// via the ShopkeeperAddedEvent.
+		if (!ShopkeepersAPI.isEnabled()) {
 			return;
 		}
 
-		ShopkeepersAPI.getShopkeeperRegistry().getAllShopkeepers().forEach(this::removeShopkeeper);
+		var allShopkeepers = ShopkeepersAPI.getShopkeeperRegistry().getAllShopkeepers();
+		allShopkeepers.forEach(shopkeeper -> this.addShopkeeper(blueMapApi, shopkeeper));
+		plugin.getLogger().info("Added BlueMap markers for all shopkeepers: "
+				+ allShopkeepers.size());
+	}
+
+	// Potentially called off the main thread. Do not access the ShopkeepersAPI here.
+	private void removeAllShopkeepersAsync(BlueMapAPI blueMapApi) {
+		// Remove the shopkeepers marker set from all maps:
+		var markerCount = 0;
+		for (var world : blueMapApi.getWorlds()) {
+			for (BlueMapMap map : world.getMaps()) {
+				var markerSet = this.removeMarkerSet(map);
+				if (markerSet != null) {
+					markerCount += markerSet.getMarkers().size();
+				}
+			}
+		}
+
+		plugin.getLogger().info("Removed " + markerCount + " BlueMap markers for all shopkeepers.");
 	}
 
 	void addShopkeeper(Shopkeeper shopkeeper) {
-		if (this.blueMapApi == null) {
-			return;
-		}
+		this.runBlueMapOperation(blueMapApi -> {
+			this.addShopkeeper(blueMapApi, shopkeeper);
+		});
+	}
+
+	private void addShopkeeper(BlueMapAPI blueMapApi, Shopkeeper shopkeeper) {
+		assert blueMapApi != null;
+		assert shopkeeper != null;
 
 		var worldName = shopkeeper.getWorldName();
 		if (worldName == null) {
@@ -222,27 +336,23 @@ public class ShopkeepersBlueMap {
 			return;
 		}
 
-		blueMapApi.getWorld(worldName).map(BlueMapWorld::getMaps).ifPresent(maps -> {
-			var shopTypeId = shopkeeper.getType().getIdentifier();
-			var markerIcon = plugin.getSettings().getMarkerIcon(shopTypeId);
-			if (markerIcon == null || markerIcon.isBlank()) {
-				return; // Skip if no marker icon is defined
-			}
+		var shopTypeId = shopkeeper.getType().getIdentifier();
+		var markerIcon = plugin.getSettings().getMarkerIcon(shopTypeId);
+		if (markerIcon == null || markerIcon.isBlank()) {
+			// Skip if no marker icon is defined:
+			plugin.debug(shopkeeper.getLogPrefix()
+					+ "Not adding BlueMap markers: No icon defined.");
+			return;
+		}
 
+		blueMapApi.getWorld(worldName).map(BlueMapWorld::getMaps).ifPresent(maps -> {
 			int anchorX = plugin.getSettings().getMarkerAnchorX(shopTypeId);
 			int anchorY = plugin.getSettings().getMarkerAnchorY(shopTypeId);
 			var markerLabel = this.getMarkerLabel(shopkeeper);
 			var detail = this.getShopkeeperDetail(shopkeeper);
 
 			for (BlueMapMap map : maps) {
-				MarkerSet markerSet = map.getMarkerSets().computeIfAbsent(
-						MARKERSET_ID,
-						key -> MarkerSet.builder()
-								.label(plugin.getSettings().getMarkerSetName())
-								.defaultHidden(false)
-								.toggleable(true)
-								.build()
-				);
+				MarkerSet markerSet = this.getOrCreateMarkerSet(map);
 
 				POIMarker marker = POIMarker.builder()
 						.label(markerLabel)
@@ -304,35 +414,46 @@ public class ShopkeepersBlueMap {
 	}
 
 	void removeShopkeeper(Shopkeeper shopkeeper) {
-		if (this.blueMapApi == null) {
-			return;
-		}
+		this.runBlueMapOperation(blueMapApi -> {
+			this.removeShopkeeper(blueMapApi, shopkeeper);
+		});
+	}
 
-		var worldName = shopkeeper.getWorldName();
-		if (worldName == null) {
-			// E.g. the case for virtual shopkeepers.
-			plugin.debug(shopkeeper.getLogPrefix()
-					+ "Not removing BlueMap markers for virtual shopkeeper.");
-			return;
-		}
+	private void removeShopkeeper(BlueMapAPI blueMapApi, Shopkeeper shopkeeper) {
+		assert blueMapApi != null;
+		assert shopkeeper != null;
 
-		blueMapApi.getWorld(worldName).map(BlueMapWorld::getMaps).ifPresent(maps -> {
-			for (BlueMapMap map : maps) {
-				MarkerSet markerSet = map.getMarkerSets().get(MARKERSET_ID);
+		// Not skipping virtual shopkeepers here: Maybe the shopkeeper object type changed in the
+		// meantime from previously non-virtual to now virtual.
+
+		var markerId = this.getMarkerId(shopkeeper);
+
+		// Check all worlds: We currently don't remember the world we previously added the marker
+		// to and we cannot use the shopkeeper's world since it might have changed since the marker
+		// was added.
+		var markerCount = 0;
+		for (var world : blueMapApi.getWorlds()) {
+			for (BlueMapMap map : world.getMaps()) {
+				MarkerSet markerSet = this.getMarkerSet(map);
 				if (markerSet == null) {
 					continue;
 				}
 
-				markerSet.getMarkers().remove(this.getMarkerId(shopkeeper));
+				var marker = markerSet.getMarkers().remove(markerId);
+				if (marker != null) {
+					markerCount += 1;
+				}
 			}
+		}
 
-			plugin.debug(shopkeeper.getLogPrefix()
-					+ "Removed BlueMap markers from " + maps.size() + " maps.");
-		});
+		plugin.debug(shopkeeper.getLogPrefix() + "Removed BlueMap markers from " + markerCount
+				+ " maps.");
 	}
 
 	void updateShopkeeper(Shopkeeper shopkeeper) {
-		this.removeShopkeeper(shopkeeper);
-		this.addShopkeeper(shopkeeper);
+		this.runBlueMapOperation(blueMapApi -> {
+			this.removeShopkeeper(blueMapApi, shopkeeper);
+			this.addShopkeeper(blueMapApi, shopkeeper);
+		});
 	}
 }